+++ src/test/java/racingcar/FakeInput.java
@@ -1,4 +1,4 @@
-package racingcar.racingGame;
+package racingcar;
 
 import racingcar.Input;
