+++ src/test/java/racingcar/RacingGameTest.java
@@ -1,9 +1,7 @@
-package racingcar.racingGame;
+package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.Input;
-import racingcar.RacingGame;
 
 import static org.assertj.core.api.Assertions.assertThat;
